neutral edit bar when data type has no parents
